requirements.txt
speedtest-cli
dash
plotly
pandas

